+++ ChangeLog.md
@@ -1,3 +1,5 @@
 # Changelog for mtlstats
 
 ## Unreleased changes
+
+- Overtime losses don't count in the loss column